4.28. Refatorando a exclusão de cozinhas para usar domain services

+++ src/main/java/com/home/algafood/infrastructure/repository/CozinhaRepositoryImpl.java
@@ -6,6 +6,7 @@ import javax.persistence.EntityManager;
 import javax.persistence.PersistenceContext;
 import javax.transaction.Transactional;
 
+import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.stereotype.Component;
 
 import com.home.algafood.domain.model.Cozinha;
@@ -13,31 +14,35 @@ import com.home.algafood.domain.repository.CozinhaRepository;
 
 @Component
 public class CozinhaRepositoryImpl implements CozinhaRepository {
-	@PersistenceContext
-	private EntityManager entityManager;
+    @PersistenceContext
+    private EntityManager entityManager;
 
-	@Override
-	public List<Cozinha> listar() {
-		return entityManager.createQuery("from Cozinha", Cozinha.class).getResultList();
+    @Override
+    public List<Cozinha> listar() {
+        return entityManager.createQuery("from Cozinha", Cozinha.class).getResultList();
 
-	}
+    }
 
-	@Override
-	public Cozinha buscar(Long id) {
-		return entityManager.find(Cozinha.class, id);
-	}
+    @Override
+    public Cozinha buscar(Long id) {
+        return entityManager.find(Cozinha.class, id);
+    }
 
-	@Transactional
-	@Override
-	public Cozinha salvar(Cozinha cozinha) {
-		return entityManager.merge(cozinha);
-	}
+    @Transactional
+    @Override
+    public Cozinha salvar(Cozinha cozinha) {
+        return entityManager.merge(cozinha);
+    }
 
-	@Transactional
-	@Override
-	public void remover(Cozinha cozinha) {
-		cozinha = buscar(cozinha.getId());
-		entityManager.remove(cozinha);
-	}
+    @Transactional
+    @Override
+    public void remover(Long id) {
+        Cozinha cozinha = buscar(id);
+
+        if (cozinha == null) {
+            throw new EmptyResultDataAccessException(1);
+        }
+        entityManager.remove(cozinha);
+    }
 
 }
